requirements.txt
transformers>=4.30.0
datasets>=2.10.0
torch>=1.12.0
accelerate>=0.18.0
sentencepiece
torchvision
pillow
streamlit
python-multipart
scikit-learn
evaluate
ftfy
rouge_score
